Move most of the I/O to the model
