+++ test/io_unit_test.yaml
@@ -1,7 +1,7 @@
 filter:
   output_filename: unit_test.h5
+  verbose: false
 
 model:
   llw2d:
     station_filename: station_test.txt
-    verbose: false
